Correct small issues in the mock documentation
Attempting to use the second example, as described, results in two
errors:

call_user_func_array(): Argument #1 ($callback) must be a valid
callback, class Mockery\Expectation does not have a method "isSet"

This is because, as described, an ExpectationInterface object is
returned, not a MockInterface object.

Method isSet(<Any Arguments>) from
Mockery_0_WeatherStation_Service_WeatherService should be called exactly
1 times but called 0 times.

This is because of using once().

I suggest the fix in this commit to correct both issues. I appreciate
it's an arbitrary example, but I feel that this will help.